Add `is_owned` and `is_owned_by` methods to `CodeOwners` struct
This commit adds two methods to the `CodeOwners` struct that allow
checking if a given user is an owner of a given path. The `is_owned`
method checks if the user is an owner of the path, while the
`is_owned_by` method checks if the user is an owner of the path and is
the only owner.