Augment `orphans` command to avoid exhausting available memory
This commit changes the command to avoid constructing the entire
codeowners Regex in memory, and using the file finder to lookup
for files that don't match the total pattern set. Whilst this is
an elegant solution, it is not the most efficient one, as it
requires for very large regex patterns to be evaluated for each
file in the repository.

Instead, we find all of the files that match the "base" pattern
and then use the `CodeOwners` struct to find the files that don't
have an assigned owner. This is a more efficient solution, and avoids
exhausting available memory.

Furthermore, we also avoid even constructing a parallel iterator if the
number of files is less than a 1000, as the overhead of creating the
thread pool and the parallel iterator is not worth it for such a small
number of files.