Simplify path matching algorithm
When loading the codeowners file, avoid keeping track of the `/`
prefix on paths.

Similarly, when scanning for paths, we always strip the base path
`prefix` from the path before matching.

This means that we can both simplify the regex matching, and
avoid having to inject `/` prefixes in the regexes when scanning
for paths across the workspace.